Moved shapeChecker to compiler
