+++ packages/core/src/compiler/shapeChecker/CheckShape.ts
